refactor: 사용하지 않는 DAO 객체 제거, dao -> repository로 변경
