+++ src/main/java/kitchenpos/application/OrderService.java
@@ -3,10 +3,6 @@ package kitchenpos.application;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import kitchenpos.dao.jpa.JpaMenuRepository;
-import kitchenpos.dao.jpa.JpaOrderLineItemRepository;
-import kitchenpos.dao.jpa.JpaOrderRepository;
-import kitchenpos.dao.jpa.JpaOrderTableRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
@@ -15,6 +11,10 @@ import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.Quantity;
 import kitchenpos.dto.request.CreateOrderLineItemRequest;
 import kitchenpos.dto.request.OrderCreateRequest;
+import kitchenpos.repository.JpaMenuRepository;
+import kitchenpos.repository.JpaOrderLineItemRepository;
+import kitchenpos.repository.JpaOrderRepository;
+import kitchenpos.repository.JpaOrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
