--- src/main/java/kitchenpos/dao/MenuDao.java
@@ -1,16 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.Menu;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface MenuDao {
-    Menu save(Menu entity);
-
-    Optional<Menu> findById(Long id);
-
-    List<Menu> findAll();
-
-    long countByIdIn(List<Long> ids);
-}
