--- src/main/java/kitchenpos/dao/MenuProductDao.java
@@ -1,16 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.MenuProduct;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface MenuProductDao {
-    MenuProduct save(MenuProduct entity);
-
-    Optional<MenuProduct> findById(Long id);
-
-    List<MenuProduct> findAll();
-
-    List<MenuProduct> findAllByMenuId(Long menuId);
-}
