+++ src/main/java/kitchenpos/repository/JpaMenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao.jpa;
+package kitchenpos.repository;
 
 import kitchenpos.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
