+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -12,7 +12,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.common.annotation.IntegrationTest;
-import kitchenpos.dao.jpa.JpaMenuGroupRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.Price;
@@ -26,6 +25,7 @@ import kitchenpos.exception.MenuGroupExceptionType;
 import kitchenpos.exception.PriceException;
 import kitchenpos.exception.ProductException;
 import kitchenpos.exception.ProductExceptionType;
+import kitchenpos.repository.JpaMenuGroupRepository;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
