+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -12,11 +12,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import java.time.LocalDateTime;
 import java.util.List;
 import kitchenpos.common.annotation.IntegrationTest;
-import kitchenpos.dao.jpa.JpaMenuGroupRepository;
-import kitchenpos.dao.jpa.JpaMenuRepository;
-import kitchenpos.dao.jpa.JpaOrderRepository;
-import kitchenpos.dao.jpa.JpaOrderTableRepository;
-import kitchenpos.dao.jpa.JpaTableGroupRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.Order;
@@ -29,6 +24,11 @@ import kitchenpos.dto.request.CreateOrderLineItemRequest;
 import kitchenpos.dto.request.OrderCreateRequest;
 import kitchenpos.exception.BaseExceptionType;
 import kitchenpos.exception.OrderException;
+import kitchenpos.repository.JpaMenuGroupRepository;
+import kitchenpos.repository.JpaMenuRepository;
+import kitchenpos.repository.JpaOrderRepository;
+import kitchenpos.repository.JpaOrderTableRepository;
+import kitchenpos.repository.JpaTableGroupRepository;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
