+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -11,8 +11,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Stream;
 import kitchenpos.common.annotation.IntegrationTest;
-import kitchenpos.dao.jpa.JpaOrderTableRepository;
-import kitchenpos.dao.jpa.JpaTableGroupRepository;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
@@ -24,6 +22,8 @@ import kitchenpos.dto.request.TableGroupCreateRequest;
 import kitchenpos.exception.BaseExceptionType;
 import kitchenpos.exception.OrderTableException;
 import kitchenpos.exception.TableGroupException;
+import kitchenpos.repository.JpaOrderTableRepository;
+import kitchenpos.repository.JpaTableGroupRepository;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
