+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -15,10 +15,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import java.time.LocalDateTime;
 import java.util.List;
 import kitchenpos.common.annotation.IntegrationTest;
-import kitchenpos.dao.jpa.JpaMenuRepository;
-import kitchenpos.dao.jpa.JpaOrderRepository;
-import kitchenpos.dao.jpa.JpaOrderTableRepository;
-import kitchenpos.dao.jpa.JpaTableGroupRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
@@ -29,6 +25,10 @@ import kitchenpos.dto.request.OrderTableCreateRequest;
 import kitchenpos.exception.BaseExceptionType;
 import kitchenpos.exception.OrderException;
 import kitchenpos.exception.OrderTableException;
+import kitchenpos.repository.JpaMenuRepository;
+import kitchenpos.repository.JpaOrderRepository;
+import kitchenpos.repository.JpaOrderTableRepository;
+import kitchenpos.repository.JpaTableGroupRepository;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
